refactor(routes): move query helper functions to utils
